+++ src/utils.test.js
@@ -3,6 +3,7 @@ import '../test/jest/__mock__';
 import { FormattedMessage } from 'react-intl';
 
 import { updateTenant } from '@folio/stripes/core';
+import { OKAPI_TENANT_HEADER } from '@folio/stripes-inventory-components';
 
 import buildStripes from '../test/jest/__mock__/stripesCore.mock';
 
@@ -11,17 +12,15 @@ import {
   validateFieldLength,
   validateNumericField,
   validateAlphaNumericField,
-  getQueryTemplate,
   switchAffiliation,
   setRecordForDeletion,
   parseEmptyFormValue,
+  redirectToMarcEditPage,
   sendCalloutOnAffiliationChange,
 } from './utils';
 import {
   CONTENT_TYPE_HEADER,
-  OKAPI_TENANT_HEADER,
   OKAPI_TOKEN_HEADER,
-  browseModeOptions,
 } from './constants';
 
 describe('validateRequiredField', () => {
@@ -121,40 +120,6 @@ describe('validateAlphaNumericField', () => {
   it('should return validation error when field is none alphaNumeric', () => {
     const expectedResult = <FormattedMessage id="ui-inventory.hridHandling.validation.assignPrefixField" />;
     expect(validateAlphaNumericField(nonAlphaNumeric)).toEqual(expectedResult);
-  });
-});
-
-describe('getQueryTemplate', () => {
-  const indexes = [
-    {
-      label: 'ui-inventory.browse.callNumbers',
-      queryTemplate: '%{query.query}',
-      subIndexes: [
-        { label: 'ui-inventory.browse.callNumbersAll', value: browseModeOptions.CALL_NUMBERS },
-        { label: 'ui-inventory.browse.dewey', value: browseModeOptions.DEWEY },
-        { label: 'ui-inventory.browse.libOfCongress', value: browseModeOptions.LIBRARY_OF_CONGRESS },
-        { label: 'ui-inventory.browse.local', value: browseModeOptions.LOCAL },
-        { label: 'ui-inventory.browse.natLibOfMed', value: browseModeOptions.NATIONAL_LIBRARY_OF_MEDICINE },
-        { label: 'ui-inventory.browse.other', value: browseModeOptions.OTHER },
-        { label: 'ui-inventory.browse.superintendent', value: browseModeOptions.SUPERINTENDENT },
-      ],
-    },
-    { label: 'ui-inventory.browse.contributors', value: browseModeOptions.CONTRIBUTORS, queryTemplate: '%{query.query}' },
-    { label: 'ui-inventory.browse.subjects', value: browseModeOptions.SUBJECTS, queryTemplate: '%{query.query}' },
-  ];
-
-  describe('when a searchable index is in sub indexes', () => {
-    it('should return a queryTemplate', () => {
-      const queryIndex = browseModeOptions.DEWEY;
-      expect(getQueryTemplate(queryIndex, indexes)).toBe('%{query.query}');
-    });
-  });
-
-  describe('when a searchable index is not in sub indexes', () => {
-    it('should return a queryTemplate', () => {
-      const queryIndex = browseModeOptions.CONTRIBUTORS;
-      expect(getQueryTemplate(queryIndex, indexes)).toBe('%{query.query}');
-    });
   });
 });
 
@@ -277,6 +242,28 @@ describe('parseEmptyFormValue', () => {
     const value = undefined;
 
     expect(parseEmptyFormValue(value)).toEqual(undefined);
+  });
+});
+
+describe('redirectToMarcEditPage', () => {
+  it('should call history.push with correct arguments', () => {
+    const pathname = 'some-pathname';
+    const instance = {
+      shared: true,
+    };
+    const location = {
+      search: '?someValue=test&relatedRecordVersion=1',
+    };
+    const history = {
+      push: jest.fn(),
+    };
+
+    redirectToMarcEditPage(pathname, instance, location, history);
+
+    expect(history.push).toHaveBeenCalledWith({
+      pathname,
+      search: 'someValue=test&shared=true',
+    });
   });
 });