+++ src/views/BrowseInventory/BrowseInventory.js
@@ -27,6 +27,7 @@ import {
   useItemToView,
   useLocationFilters,
 } from '@folio/stripes-acq-components';
+import { useResetFacetStates } from '@folio/stripes-inventory-components';
 
 import {
   BrowseInventoryFilters,
@@ -47,6 +48,7 @@ const BrowseInventory = () => {
   const location = useLocation();
   const intl = useIntl();
   const [namespace] = useNamespace();
+  const resetFacetStates = useResetFacetStates();
   const {
     getLastSearch,
     getLastBrowseOffset,
@@ -157,6 +159,7 @@ const BrowseInventory = () => {
   }, [searchQuery, filters]);
 
   const onChangeSearchIndex = useCallback((e) => {
+    resetFacetStates({ isBrowseLookup: true });
     deleteItemToView();
     /*
       useLocationFilters hook returns `resetLocationFilters` function, but it also resets search index and query, which we want to avoid in our case
@@ -165,12 +168,13 @@ const BrowseInventory = () => {
     */
     changeSearchIndex(e);
     clearFilters();
-  }, [deleteItemToView, clearFilters, changeSearchIndex]);
+  }, [deleteItemToView, clearFilters, changeSearchIndex, resetFacetStates]);
 
   const onReset = useCallback(() => {
+    resetFacetStates({ isBrowseLookup: true });
     resetAll();
     inputRef.current.focus();
-  }, [inputRef.current, resetAll]);
+  }, [inputRef.current, resetAll, resetFacetStates]);
 
   useEffect(() => {
     if (hasFocusedSearchOptionOnMount.current) {
@@ -221,9 +225,8 @@ const BrowseInventory = () => {
           />
 
           <BrowseInventoryFilters
-            activeFilters={filters}
+            query={filters}
             applyFilters={applyFilters}
-            searchIndex={searchIndex}
           />
         </FiltersPane>
       )}